requirements.txt
Django==5.0.6
django-environ==0.11.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
argon2-cffi==23.1.0
setuptools==75.6.0